refactor(indexer): indexer theme files
