+++ applications/tari_indexer_web_ui/src/assets/Logo.tsx
@@ -1,0 +1,97 @@
+//  Copyright 2022. The Tari Project
+//
+//  Redistribution and use in source and binary forms, with or without modification, are permitted provided that the
+//  following conditions are met:
+//
+//  1. Redistributions of source code must retain the above copyright notice, this list of conditions and the following
+//  disclaimer.
+//
+//  2. Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the
+//  following disclaimer in the documentation and/or other materials provided with the distribution.
+//
+//  3. Neither the name of the copyright holder nor the names of its contributors may be used to endorse or promote
+//  products derived from this software without specific prior written permission.
+//
+//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES,
+//  INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+//  DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+//  SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+//  SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+//  WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+//  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+interface LogoProps {
+  width?: string;
+  height?: string;
+  fill?: string;
+}
+
+const Logo: React.FC<LogoProps> = ({
+  width = '100%',
+  height = 'auto',
+  fill = 'black',
+}) => (
+  <svg
+    width={width}
+    height={height}
+    viewBox="0 -5 350 60"
+    xmlns="http://www.w3.org/2000/svg"
+  >
+    <path
+      d="M197.896 8.008H204.592C205.712 8.008 206.792 8.184 207.832 8.536C208.872 8.872 209.792 9.392 210.592 10.096C211.392 10.8 212.032 11.688 212.512 12.76C212.992 13.816 213.232 15.064 213.232 16.504C213.232 17.96 212.952 19.224 212.392 20.296C211.848 21.352 211.136 22.232 210.256 22.936C209.392 23.624 208.424 24.144 207.352 24.496C206.296 24.832 205.256 25 204.232 25H197.896V8.008ZM203.176 22.264C204.12 22.264 205.008 22.16 205.84 21.952C206.688 21.728 207.424 21.392 208.048 20.944C208.672 20.48 209.16 19.888 209.512 19.168C209.88 18.432 210.064 17.544 210.064 16.504C210.064 15.48 209.904 14.6 209.584 13.864C209.264 13.128 208.816 12.536 208.24 12.088C207.68 11.624 207.008 11.288 206.224 11.08C205.456 10.856 204.608 10.744 203.68 10.744H200.92V22.264H203.176Z"
+      fill={fill}
+    />
+    <path
+      d="M221.603 8.008H224.219L231.539 25H228.083L226.499 21.112H219.131L217.595 25H214.211L221.603 8.008ZM225.443 18.52L222.827 11.608L220.163 18.52H225.443Z"
+      fill={fill}
+    />
+    <path
+      d="M233.427 8.008H237.435L245.667 20.632H245.715V8.008H248.739V25H244.899L236.499 11.968H236.451V25H233.427V8.008Z"
+      fill={fill}
+    />
+    <path d="M197.824 30.008H200.848V47H197.824V30.008Z" fill={fill} />
+    <path
+      d="M204.576 30.008H208.584L216.816 42.632H216.864V30.008H219.888V47H216.048L207.648 33.968H207.6V47H204.576V30.008Z"
+      fill={fill}
+    />
+    <path
+      d="M223.677 30.008H230.373C231.493 30.008 232.573 30.184 233.613 30.536C234.653 30.872 235.573 31.392 236.373 32.096C237.173 32.8 237.813 33.688 238.293 34.76C238.773 35.816 239.013 37.064 239.013 38.504C239.013 39.96 238.733 41.224 238.173 42.296C237.629 43.352 236.917 44.232 236.037 44.936C235.173 45.624 234.205 46.144 233.133 46.496C232.077 46.832 231.037 47 230.013 47H223.677V30.008ZM228.957 44.264C229.901 44.264 230.789 44.16 231.621 43.952C232.469 43.728 233.205 43.392 233.829 42.944C234.453 42.48 234.941 41.888 235.293 41.168C235.661 40.432 235.845 39.544 235.845 38.504C235.845 37.48 235.685 36.6 235.365 35.864C235.045 35.128 234.597 34.536 234.021 34.088C233.461 33.624 232.789 33.288 232.005 33.08C231.237 32.856 230.389 32.744 229.461 32.744H226.701V44.264H228.957Z"
+      fill={fill}
+    />
+    <path
+      d="M241.888 30.008H253.144V32.744H244.912V36.92H252.712V39.656H244.912V44.264H253.576V47H241.888V30.008Z"
+      fill={fill}
+    />
+    <path
+      d="M260.712 38.12L255.048 30.008H258.864L262.872 36.296L266.832 30.008H270.504L264.936 38.12L271.104 47H267.264L262.776 39.824L258.312 47H254.664L260.712 38.12Z"
+      fill={fill}
+    />
+    <path
+      d="M272.99 30.008H284.246V32.744H276.014V36.92H283.814V39.656H276.014V44.264H284.678V47H272.99V30.008Z"
+      fill={fill}
+    />
+    <path
+      d="M287.662 30.008H293.566C294.382 30.008 295.166 30.088 295.918 30.248C296.686 30.392 297.366 30.648 297.958 31.016C298.55 31.384 299.022 31.88 299.374 32.504C299.726 33.128 299.902 33.92 299.902 34.88C299.902 36.112 299.558 37.144 298.87 37.976C298.198 38.808 297.23 39.312 295.966 39.488L300.478 47H296.83L292.894 39.8H290.686V47H287.662V30.008ZM293.038 37.208C293.47 37.208 293.902 37.192 294.334 37.16C294.766 37.112 295.158 37.016 295.51 36.872C295.878 36.712 296.174 36.48 296.398 36.176C296.622 35.856 296.734 35.416 296.734 34.856C296.734 34.36 296.63 33.96 296.422 33.656C296.214 33.352 295.942 33.128 295.606 32.984C295.27 32.824 294.894 32.72 294.478 32.672C294.078 32.624 293.686 32.6 293.302 32.6H290.686V37.208H293.038Z"
+      fill={fill}
+    />
+    <path
+      d="M115.148 30.8698L118.671 19.301L122.194 30.8698H115.148ZM114.869 8.29834L101.728 47.1145H110.134L112.908 38.14H124.434L127.208 47.1145H135.614L122.472 8.29834H114.869Z"
+      fill={fill}
+    />
+    <path d="M176.445 8H184.258V46.8162H176.445V8Z" fill={fill} />
+    <path
+      d="M147.87 24.6427V15.2703H154.515C157.546 15.2703 159.147 16.8907 159.147 19.9569C159.147 23.0225 157.546 24.6427 154.515 24.6427H147.87ZM155.777 31.8769C162.953 31.5006 167.069 27.156 167.069 19.9569C167.069 12.4699 162.477 8 154.787 8H140.057V46.8159H147.87V34.6857L158.571 46.8159H168.634L155.354 31.8991L155.777 31.8769Z"
+      fill={fill}
+    />
+    <path
+      d="M82.5313 47.1145H90.3446V15.5689H103.461V8.29834H69.4141V15.5689H82.5313V47.1145Z"
+      fill={fill}
+    />
+    <path
+      d="M49.489 18.4517L49.4826 25.1813L9.9661 15.0154L23.3114 6.31705L49.489 18.4517ZM25.5188 46.7641L25.5086 24.732L46.5851 30.1605L25.5188 46.7641ZM20.0033 44.6529L5.51421 28.4097L5.50562 19.5436L19.9803 23.3081L20.0033 44.6529ZM0 14.9387L0.00212257 30.5344L22.7523 56L54.9548 30.6016L55 14.9151L22.826 0L0 14.9387Z"
+      fill={fill}
+    />
+  </svg>
+);
+
+export default Logo;
