--- applications/tari_indexer_web_ui/src/index.css
@@ -1,40 +1,0 @@
-@font-face {
-  src: url('./assets/fonts/AvenirLTStd-Book.otf');
-  font-family: 'AvenirRegular'
-}
-
-@font-face {
-  src: url('./assets/fonts/AvenirLTStd-Medium.otf');
-  font-family: 'AvenirMedium'
-}
-
-@font-face {
-  src: url('./assets/fonts/AvenirLTStd-Heavy.otf');
-  font-family: 'AvenirHeavy'
-}
-
-body {
-  margin: 0;
-  font-family: 'AvenirMedium', sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  background-color: #F5F5F7;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
-}
-
-.add-confirm-form {
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  gap: 10px;
-}
-
-@media screen and (min-width: 768px) {
-  
-.add-confirm-form {
-  flex-direction: row;
-}
-}
